new permissionModule

+++ permission-library/src/main/java/android/support/v4ox/widget/NestedScrollView.java
@@ -666,7 +666,7 @@ public class NestedScrollView extends FrameLayout implements NestedScrollingPare
 
             case MotionEvent.ACTION_DOWN: {
                 final int y = (int) ev.getY();
-                if (!inChild((int) ev.getX(), (int) y)) {
+                if (!inChild((int) ev.getX(), y)) {
                     mIsBeingDragged = false;
                     recycleVelocityTracker();
                     break;
